+++ packages/compiler/src/transforms/lazy-state.ts
@@ -11,6 +11,32 @@ import type { NodePath } from 'ast-types/lib/node-path';
 import type { Transformer } from '../transform';
 import IgnoreRules from '../transform-ignore';
 
+type LazyStateErrorOptions = {
+  details?: string;
+  fix?: string;
+  pos?: any;
+};
+
+export class LazyStateError extends Error {
+  fix?: string;
+  details?: string;
+  pos?: any;
+
+  constructor(
+    message: string,
+    { details, fix, pos }: LazyStateErrorOptions = {}
+  ) {
+    const posStr = (pos && `(${pos.start.line}:${pos.start.column})`) ?? '';
+    super(`Lazy State Error: ${message} ${posStr}`);
+
+    this.fix = fix;
+    this.details = details;
+    const { start, end } = pos;
+    this.pos = { start, end };
+    delete this.stack;
+  }
+}
+
 // Walk up the AST and work out where the parent arrow function should go
 const ensureParentArrow = (path: NodePath<n.MemberExpression>) => {
   let root = path;
@@ -23,9 +49,12 @@ const ensureParentArrow = (path: NodePath<n.MemberExpression>) => {
     root = root.parent;
 
     // if this is any kind of statement, we should throw
-    // TODO we may relax this, see https://github.com/OpenFn/kit/issues/660
     if (n.Statement.check(root.node) || n.Declaration.check(root.node)) {
-      throw new Error(`invalid state operator: must be inside an expression`);
+      throw new LazyStateError('Must be inside an operation', {
+        pos: path.node.loc,
+        details:
+          'The Lazy State operation must be used inside a top-level operation, like fn(). It cannot be used inside a regular JavaScript statement because no valid state reference is available.',
+      });
     }
   }
 
@@ -47,9 +76,9 @@ const ensureParentArrow = (path: NodePath<n.MemberExpression>) => {
     }
   } else {
     // Actually I don't think we'll ever get here
-    throw new Error(
-      `invalid state operator: must be be passed as an argument to an operator`
-    );
+    throw new LazyStateError('must be passed as an argument to an operator', {
+      pos: path.node.loc,
+    });
   }
 };
 
@@ -66,11 +95,11 @@ const isOpenFunction = (path: NodePath) => {
         // We already have a valid open function here
         return true;
       }
-      throw new Error(
-        `invalid state operator: parameter "${name}" should be called "state"`
-      );
+      throw new LazyStateError(`parameter "${name}" should be called "state"`, {
+        pos: path.node.loc,
+      });
     }
-    throw new Error('invalid state operator: parent has wrong arity');
+    throw new LazyStateError('parent has wrong arity', { pos: path.node.loc });
   }
 
   // if we get here, then path is: